refactor(artifacts): define prepared download identities and formats
