+++ src/notebooklm/types.py
@@ -22,6 +22,7 @@ from ._types.artifact_content import (
     FlashcardArtifactUserState,
     UnknownArtifactUserState,
 )
+from ._types.artifact_download import ArtifactDownloadRequest, ArtifactDownloadSelection
 from ._types.artifacts import (
     Artifact,
     ArtifactCustomizationChoices,
@@ -220,6 +221,8 @@ _CITED_SOURCE_SELECTION_TYPE_HINT_GLOBALS = (ResearchSourceInput,)
 
 
 __all__ = [
+    "ArtifactDownloadRequest",
+    "ArtifactDownloadSelection",
     "SourceDeleteOutcome",
     # Dataclasses
     "AccountLimits",
@@ -376,6 +379,8 @@ del _public_common_type
 
 
 for _public_moved_type in (
+    ArtifactDownloadRequest,
+    ArtifactDownloadSelection,
     Artifact,
     ArtifactCustomizationChoices,
     ArtifactInfographic,
